refactor: 스토리북 전역 provider decorator preview에서 적용
