+++ src/components/CartController/CartController.stories.tsx
@@ -1,17 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import CartController from './CartController';
-import Providers from '../../stories/Providers';
 
 const meta = {
   component: CartController,
   title: 'CartController',
-  decorators: [
-    (Story) => (
-      <Providers>
-        <Story />
-      </Providers>
-    ),
-  ],
 } satisfies Meta<typeof CartController>;
 
 export default meta;
