+++ src/components/Header/Header.stories.tsx
@@ -1,17 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import Header from '.';
-import Providers from '../../stories/Providers';
 
 const meta = {
   component: Header,
   title: 'Header',
-  decorators: [
-    (Story) => (
-      <Providers>
-        <Story />
-      </Providers>
-    ),
-  ],
 } satisfies Meta<typeof Header>;
 
 export default meta;
